Implement jar discovery adapters.
By registering ILabelJar adapters for Interface and IApplication
we can make the ILabelJar-adapter work for any object.
If the object does not implement ILabelRoot, the discovery adapter
kicks in and walks up the acquisition chain recursively until the
root is found and the ILabelJar adapter for this scope is returned.

If the Zope application root is reached, a LookupError is raised.