Fix redirecting and tests.
The referer in the overlay is actually incorrect, because it is not
embedded in a frame and therefore actually uses the referer of the
parent page.

The testbrowser does not support overlays (nor JavaScript) and
therefore directly tested the view, using different referers.

Also the fallback to the absolute url did not work because the
request's get method does not use the default value as expected.